feat(wallet)!: return routerV2 suggested routes whenever available, on top of error if the best route doesn't pass all checks
